fs: refactor fs module
PR-URL: https://github.com/nodejs/node/pull/20764
