+++ lib/internal/fs/streams.js
@@ -1,0 +1,362 @@
+'use strict';
+
+const {
+  FSReqWrap,
+  writeBuffers
+} = process.binding('fs');
+const {
+  ERR_INVALID_ARG_TYPE,
+  ERR_OUT_OF_RANGE
+} = require('internal/errors').codes;
+const fs = require('fs');
+const { Buffer } = require('buffer');
+const {
+  copyObject,
+  getOptions,
+} = require('internal/fs/utils');
+const { Readable, Writable } = require('stream');
+const { getPathFromURL } = require('internal/url');
+const util = require('util');
+
+const kMinPoolSpace = 128;
+
+let pool;
+
+function allocNewPool(poolSize) {
+  pool = Buffer.allocUnsafe(poolSize);
+  pool.used = 0;
+}
+
+function ReadStream(path, options) {
+  if (!(this instanceof ReadStream))
+    return new ReadStream(path, options);
+
+  // a little bit bigger buffer and water marks by default
+  options = copyObject(getOptions(options, {}));
+  if (options.highWaterMark === undefined)
+    options.highWaterMark = 64 * 1024;
+
+  // for backwards compat do not emit close on destroy.
+  options.emitClose = false;
+
+  Readable.call(this, options);
+
+  // path will be ignored when fd is specified, so it can be falsy
+  this.path = getPathFromURL(path);
+  this.fd = options.fd === undefined ? null : options.fd;
+  this.flags = options.flags === undefined ? 'r' : options.flags;
+  this.mode = options.mode === undefined ? 0o666 : options.mode;
+
+  this.start = options.start;
+  this.end = options.end;
+  this.autoClose = options.autoClose === undefined ? true : options.autoClose;
+  this.pos = undefined;
+  this.bytesRead = 0;
+  this.closed = false;
+
+  if (this.start !== undefined) {
+    if (typeof this.start !== 'number' || Number.isNaN(this.start)) {
+      throw new ERR_INVALID_ARG_TYPE('start', 'number', this.start);
+    }
+    if (this.end === undefined) {
+      this.end = Infinity;
+    } else if (typeof this.end !== 'number' || Number.isNaN(this.end)) {
+      throw new ERR_INVALID_ARG_TYPE('end', 'number', this.end);
+    }
+
+    if (this.start > this.end) {
+      const errVal = `{start: ${this.start}, end: ${this.end}}`;
+      throw new ERR_OUT_OF_RANGE('start', '<= "end"', errVal);
+    }
+
+    this.pos = this.start;
+  }
+
+  // Backwards compatibility: Make sure `end` is a number regardless of `start`.
+  // TODO(addaleax): Make the above typecheck not depend on `start` instead.
+  // (That is a semver-major change).
+  if (typeof this.end !== 'number')
+    this.end = Infinity;
+  else if (Number.isNaN(this.end))
+    throw new ERR_INVALID_ARG_TYPE('end', 'number', this.end);
+
+  if (typeof this.fd !== 'number')
+    this.open();
+
+  this.on('end', function() {
+    if (this.autoClose) {
+      this.destroy();
+    }
+  });
+}
+util.inherits(ReadStream, Readable);
+
+ReadStream.prototype.open = function() {
+  fs.open(this.path, this.flags, this.mode, (er, fd) => {
+    if (er) {
+      if (this.autoClose) {
+        this.destroy();
+      }
+      this.emit('error', er);
+      return;
+    }
+
+    this.fd = fd;
+    this.emit('open', fd);
+    this.emit('ready');
+    // start the flow of data.
+    this.read();
+  });
+};
+
+ReadStream.prototype._read = function(n) {
+  if (typeof this.fd !== 'number') {
+    return this.once('open', function() {
+      this._read(n);
+    });
+  }
+
+  if (this.destroyed)
+    return;
+
+  if (!pool || pool.length - pool.used < kMinPoolSpace) {
+    // discard the old pool.
+    allocNewPool(this.readableHighWaterMark);
+  }
+
+  // Grab another reference to the pool in the case that while we're
+  // in the thread pool another read() finishes up the pool, and
+  // allocates a new one.
+  const thisPool = pool;
+  let toRead = Math.min(pool.length - pool.used, n);
+  const start = pool.used;
+
+  if (this.pos !== undefined)
+    toRead = Math.min(this.end - this.pos + 1, toRead);
+  else
+    toRead = Math.min(this.end - this.bytesRead + 1, toRead);
+
+  // already read everything we were supposed to read!
+  // treat as EOF.
+  if (toRead <= 0)
+    return this.push(null);
+
+  // the actual read.
+  fs.read(this.fd, pool, pool.used, toRead, this.pos, (er, bytesRead) => {
+    if (er) {
+      if (this.autoClose) {
+        this.destroy();
+      }
+      this.emit('error', er);
+    } else {
+      let b = null;
+      if (bytesRead > 0) {
+        this.bytesRead += bytesRead;
+        b = thisPool.slice(start, start + bytesRead);
+      }
+
+      this.push(b);
+    }
+  });
+
+  // move the pool positions, and internal position for reading.
+  if (this.pos !== undefined)
+    this.pos += toRead;
+  pool.used += toRead;
+};
+
+ReadStream.prototype._destroy = function(err, cb) {
+  const isOpen = typeof this.fd !== 'number';
+  if (isOpen) {
+    this.once('open', closeFsStream.bind(null, this, cb, err));
+    return;
+  }
+
+  closeFsStream(this, cb, err);
+  this.fd = null;
+};
+
+function closeFsStream(stream, cb, err) {
+  fs.close(stream.fd, (er) => {
+    er = er || err;
+    cb(er);
+    stream.closed = true;
+    if (!er)
+      stream.emit('close');
+  });
+}
+
+ReadStream.prototype.close = function(cb) {
+  this.destroy(null, cb);
+};
+
+function WriteStream(path, options) {
+  if (!(this instanceof WriteStream))
+    return new WriteStream(path, options);
+
+  options = copyObject(getOptions(options, {}));
+
+  // for backwards compat do not emit close on destroy.
+  options.emitClose = false;
+
+  Writable.call(this, options);
+
+  // path will be ignored when fd is specified, so it can be falsy
+  this.path = getPathFromURL(path);
+  this.fd = options.fd === undefined ? null : options.fd;
+  this.flags = options.flags === undefined ? 'w' : options.flags;
+  this.mode = options.mode === undefined ? 0o666 : options.mode;
+
+  this.start = options.start;
+  this.autoClose = options.autoClose === undefined ? true : !!options.autoClose;
+  this.pos = undefined;
+  this.bytesWritten = 0;
+  this.closed = false;
+
+  if (this.start !== undefined) {
+    if (typeof this.start !== 'number') {
+      throw new ERR_INVALID_ARG_TYPE('start', 'number', this.start);
+    }
+    if (this.start < 0) {
+      const errVal = `{start: ${this.start}}`;
+      throw new ERR_OUT_OF_RANGE('start', '>= 0', errVal);
+    }
+
+    this.pos = this.start;
+  }
+
+  if (options.encoding)
+    this.setDefaultEncoding(options.encoding);
+
+  if (typeof this.fd !== 'number')
+    this.open();
+}
+util.inherits(WriteStream, Writable);
+
+WriteStream.prototype._final = function(callback) {
+  if (this.autoClose) {
+    this.destroy();
+  }
+
+  callback();
+};
+
+WriteStream.prototype.open = function() {
+  fs.open(this.path, this.flags, this.mode, (er, fd) => {
+    if (er) {
+      if (this.autoClose) {
+        this.destroy();
+      }
+      this.emit('error', er);
+      return;
+    }
+
+    this.fd = fd;
+    this.emit('open', fd);
+    this.emit('ready');
+  });
+};
+
+
+WriteStream.prototype._write = function(data, encoding, cb) {
+  if (!(data instanceof Buffer)) {
+    const err = new ERR_INVALID_ARG_TYPE('data', 'Buffer', data);
+    return this.emit('error', err);
+  }
+
+  if (typeof this.fd !== 'number') {
+    return this.once('open', function() {
+      this._write(data, encoding, cb);
+    });
+  }
+
+  fs.write(this.fd, data, 0, data.length, this.pos, (er, bytes) => {
+    if (er) {
+      if (this.autoClose) {
+        this.destroy();
+      }
+      return cb(er);
+    }
+    this.bytesWritten += bytes;
+    cb();
+  });
+
+  if (this.pos !== undefined)
+    this.pos += data.length;
+};
+
+
+function writev(fd, chunks, position, callback) {
+  function wrapper(err, written) {
+    // Retain a reference to chunks so that they can't be GC'ed too soon.
+    callback(err, written || 0, chunks);
+  }
+
+  const req = new FSReqWrap();
+  req.oncomplete = wrapper;
+  writeBuffers(fd, chunks, position, req);
+}
+
+
+WriteStream.prototype._writev = function(data, cb) {
+  if (typeof this.fd !== 'number') {
+    return this.once('open', function() {
+      this._writev(data, cb);
+    });
+  }
+
+  const self = this;
+  const len = data.length;
+  const chunks = new Array(len);
+  let size = 0;
+
+  for (var i = 0; i < len; i++) {
+    const chunk = data[i].chunk;
+
+    chunks[i] = chunk;
+    size += chunk.length;
+  }
+
+  writev(this.fd, chunks, this.pos, function(er, bytes) {
+    if (er) {
+      self.destroy();
+      return cb(er);
+    }
+    self.bytesWritten += bytes;
+    cb();
+  });
+
+  if (this.pos !== undefined)
+    this.pos += size;
+};
+
+
+WriteStream.prototype._destroy = ReadStream.prototype._destroy;
+WriteStream.prototype.close = function(cb) {
+  if (cb) {
+    if (this.closed) {
+      process.nextTick(cb);
+      return;
+    } else {
+      this.on('close', cb);
+    }
+  }
+
+  // If we are not autoClosing, we should call
+  // destroy on 'finish'.
+  if (!this.autoClose) {
+    this.on('finish', this.destroy.bind(this));
+  }
+
+  // we use end() instead of destroy() because of
+  // https://github.com/nodejs/node/issues/2006
+  this.end();
+};
+
+// There is no shutdown() for files.
+WriteStream.prototype.destroySoon = WriteStream.prototype.end;
+
+module.exports = {
+  ReadStream,
+  WriteStream
+};
