+++ lib/internal/fs/utils.js
@@ -1,7 +1,6 @@
 'use strict';
 
 const { Buffer, kMaxLength } = require('buffer');
-const { Writable } = require('stream');
 const {
   ERR_FS_INVALID_SYMLINK_TYPE,
   ERR_INVALID_ARG_TYPE,
@@ -11,7 +10,6 @@ const {
   ERR_OUT_OF_RANGE
 } = require('internal/errors').codes;
 const { isUint8Array } = require('internal/util/types');
-const fs = require('fs');
 const pathModule = require('path');
 const util = require('util');
 
@@ -256,45 +254,6 @@ function stringToSymlinkType(type) {
   return flags;
 }
 
-// Temporary hack for process.stdout and process.stderr when piped to files.
-function SyncWriteStream(fd, options) {
-  Writable.call(this);
-
-  options = options || {};
-
-  this.fd = fd;
-  this.readable = false;
-  this.autoClose = options.autoClose === undefined ? true : options.autoClose;
-
-  this.on('end', () => this._destroy());
-}
-
-util.inherits(SyncWriteStream, Writable);
-
-SyncWriteStream.prototype._write = function(chunk, encoding, cb) {
-  fs.writeSync(this.fd, chunk, 0, chunk.length);
-  cb();
-  return true;
-};
-
-SyncWriteStream.prototype._destroy = function() {
-  if (this.fd === null) // already destroy()ed
-    return;
-
-  if (this.autoClose)
-    fs.closeSync(this.fd);
-
-  this.fd = null;
-  return true;
-};
-
-SyncWriteStream.prototype.destroySoon =
-SyncWriteStream.prototype.destroy = function() {
-  this._destroy();
-  this.emit('close');
-  return true;
-};
-
 // converts Date or number to a fractional UNIX timestamp
 function toUnixTimestamp(time, name = 'time') {
   // eslint-disable-next-line eqeqeq
@@ -383,7 +342,6 @@ module.exports = {
   stringToFlags,
   stringToSymlinkType,
   Stats,
-  SyncWriteStream,
   toUnixTimestamp,
   validateBuffer,
   validateOffsetLengthRead,
