+++ lib/internal/process/stdio.js
@@ -167,8 +167,8 @@ function createWritableStdioStream(fd) {
       break;
 
     case 'FILE':
-      var fs = require('internal/fs/utils');
-      stream = new fs.SyncWriteStream(fd, { autoClose: false });
+      const SyncWriteStream = require('internal/fs/sync_write_stream');
+      stream = new SyncWriteStream(fd, { autoClose: false });
       stream._type = 'fs';
       break;
